Добавлен бар здоровья, уровня, защиты башни

+++ node/src/main/kotlin/me/reidj/tower/mob/Mob.kt
@@ -15,7 +15,7 @@ data class Mob(
     val x: Double,
     val y: Double,
     val z: Double,
-    val damage: Double,
+    val damage: Int,
     val type: EntityType
 ) {
